+++ src/Type/Passport/PassportElementErrorDataField.php
@@ -1,0 +1,35 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Type\Passport;
+
+/**
+ * @see https://core.telegram.org/bots/api#passportelementerrordatafield
+ */
+final readonly class PassportElementErrorDataField implements PassportElementError
+{
+    public function __construct(
+        public string $type,
+        public string $fieldName,
+        public string $dataHash,
+        public string $message,
+    ) {
+    }
+
+    public function getSource(): string
+    {
+        return 'data';
+    }
+
+    public function toRequestArray(): array
+    {
+        return [
+            'source' => $this->getSource(),
+            'type' => $this->type,
+            'field_name' => $this->fieldName,
+            'data_hash' => $this->dataHash,
+            'message' => $this->message,
+        ];
+    }
+}