+++ tests/Type/Passport/PassportElementErrorDataFieldTest.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Tests\Type\Passport;
+
+use PHPUnit\Framework\TestCase;
+use Vjik\TelegramBot\Api\Type\Passport\PassportElementErrorDataField;
+
+final class PassportElementErrorDataFieldTest extends TestCase
+{
+    public function testBase(): void
+    {
+        $type = new PassportElementErrorDataField('driver_license', 'test.jpg', 'qwerty', 'Test message');
+
+        $this->assertSame('data', $type->getSource());
+        $this->assertSame(
+            [
+                'source' => 'data',
+                'type' => 'driver_license',
+                'field_name' => 'test.jpg',
+                'data_hash' => 'qwerty',
+                'message' => 'Test message',
+            ],
+            $type->toRequestArray(),
+        );
+    }
+}