+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 - New #82: Add `createInvoiceLink` method.
 - New #83: Add `answerShippingQuery` and `answerPreCheckoutQuery` methods, and `ShippingOption` type.
 - New #84: Add `refundStarPayment` method.
+- New #85: Add `PassportElementError*` types.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.